Detecting duplicated components (#121)
Add check to detect duplicated components of a record when parsing a
TRLC file.
The Language Reference Manual has been updated accordingly, and its
version has been increased to 3.1

Example:
```
Something duplicated
{
	description = "This is fine!"
	description = "This is the duplicate."
}
```

The behavior of the `Record_Object` class is changed such that the
`assign` method only assigns components to the record if it has not been
assigned already. This is implemented by checking if the field type is
different from `Implicit_Null`.

The `Parser` class asks the `Record_Object` if the component is
`Implicit_Null`. If no, then an error is sent to the message handler.