Fix dump output for Decimal_Literal. (#140)
String formatting in trlc/ast.py is done with the old modulo formatting
syntax.

This lead to a bug, where dump() output for Decimal_Literals was
incorrect.
The value of the decimal was formated into a "%u", an unsigned int
format specifier.

Reformatted all dump() implementations in the file to use fstrings.

Co-authored-by: mugdhadhole1 <161293900+mugdhadhole1@users.noreply.github.com>